programs/otel-collector-gke/requirements.txt
pulumi>=3.110.0
pulumi-gcp>=8.40.0
pulumi-kubernetes>=4.23.0
pyyaml>=6.0.1
